pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-summarizer"
version = "0.1.0"
description = "FastAPI service that summarizes GitHub repositories from prioritized AST skeletons."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "tiktoken>=0.7.0",
  "tree-sitter>=0.23.0",
  "tree-sitter-python>=0.23.0",
  "tree-sitter-javascript>=0.23.0",
  "tree-sitter-go>=0.23.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[project.scripts]
repo-summarizer = "repo_summarizer.main:run"

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["repo_summarizer*"]
